Don't return the first object as common when the parameter is nil
When the array to compare with is nil, indexOfObject: returns nil,
which is different from NSNotFound, causing the first object to
be returned as common. This commit fixes this problem and also
avoids iterating the self array when the array to compare with
is empty (since indexOfObject: will be NSNotFound for any
object).